requirements.txt
numpy
glfw
PyOpenGL